weeks/week-17/events-s18/events-svc/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.30
psycopg2-binary>=2.9.9
grpcio>=1.64.0
grpcio-tools>=1.64.0
protobuf>=5.27.0
pydantic>=2.7.0

